+++ constants/posts.constants.js
@@ -1,0 +1,5 @@
+export const postsConstants = {
+    POST_REQUEST: 'POSTS_POST_REQUEST',
+    POST_SUCCESS: 'POSTS_POST_SUCCESS',
+    POST_FAILURE: 'POSTS_POST_FAILURE',
+};